% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{ascii.line}
\alias{ascii.line}
\title{ascii.line}
\usage{
ascii.line(
  strs,
  out.length = 66,
  left = "## ",
  right = " ##",
  sep = " ",
  justification = "center"
)
}
\arguments{
\item{justification}{}
}
\value{

}
\description{
ascii.line
}
