% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-date.R
\name{asDateTime}
\alias{asDateTime}
\title{asDateTime}
\usage{
asDateTime(strvec, from = "\%Y-\%m-\%d \%H:\%M:\%S", to = "", num = TRUE)
}
\arguments{
\item{num}{}
}
\value{

}
\description{
asDateTime
}
