% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{assignColumnsTypeInDataFrame}
\alias{assignColumnsTypeInDataFrame}
\title{assignColumnsTypeInDataFrame}
\usage{
assignColumnsTypeInDataFrame(cols, type, df)
}
\arguments{
\item{cols}{Names of columns to assign}

\item{type}{Assignment type:  "numeric" or "integer" or "character"}

\item{df}{dataframe}
}
\value{
dataframe updated
}
\description{
assignColumnsTypeInDataFrame
}
\examples{
library(datasets);
data(iris);
mydf = iris; head(mydf);
mycols = c("Sepal.Width","Petal.Width");
mydf = assignColumnsTypeInDataFrame(mycols, "integer", mydf);
head(mydf);
}
