% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-bits.R
\name{bitOr}
\alias{bitOr}
\title{bitOr}
\usage{
bitOr(a, b)
}
\arguments{
\item{a}{integer 'a'}

\item{b}{integer 'b'}
}
\value{

}
\description{
This updates the built-in functions to allow for negative integers.
Used for manual '.md5' computation, and has some issues.
Maybe R::CRAN will fix this someday in the base?
}
\details{
Specifically, this addresses overflows ...
}
\examples{
bitOr(15, 7);
}
