% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-bits.R
\name{bitShiftL}
\alias{bitShiftL}
\title{bitShiftL}
\usage{
bitShiftL(x, bits, unsigned = FALSE)
}
\arguments{
\item{x}{integer}

\item{bits}{bits}

\item{unsigned}{is it signed or not}
}
\value{
new integer
}
\description{
This updates the built-in functions to allow for negative integers.
Used for manual '.md5' computation, and has some issues.
Maybe R::CRAN will fix this someday in the base?
}
\examples{
bitShiftL(1732, 16);
bitShiftL(-1732, 16);
}
