% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-bits.R
\name{bitShiftR}
\alias{bitShiftR}
\title{bitShiftR}
\usage{
bitShiftR(x, bits, unsigned = FALSE)
}
\arguments{
\item{x}{integer}

\item{bits}{bits}

\item{unsigned}{is it signed or not}
}
\value{
new integer
}
\description{
This updates the built-in functions to allow for negative integers.
Used for manual '.md5' computation, and has some issues.
Maybe R::CRAN will fix this someday in the base?
}
\examples{
bitShiftR(1732584193, 16);
bitShiftR(-1732584193, 16);
}
