% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{buildNumericalDataForIntegral}
\alias{buildNumericalDataForIntegral}
\title{buildNumericalDataForIntegral}
\usage{
buildNumericalDataForIntegral(
  fprep = "normal: -1, 1",
  i.lim = c(NA, NA),
  x.domain = c(-10, 10),
  dxi = 0.01,
  forceEven = TRUE
)
}
\arguments{
\item{forceEven}{}
}
\value{

}
\description{
buildNumericalDataForIntegral
}
