% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{castStringAsFunction}
\alias{castStringAsFunction}
\title{castStringAsFunction}
\usage{
castStringAsFunction(fstr, ..., envir = parent.frame())
}
\arguments{
\item{fstr}{The RHS (right hand side) of a function in string form.}

\item{...}{The elements in RHS that are parameters (e.g., x)}

\item{envir}{The scope of the environment for the function}
}
\value{
A function
}
\description{
castStringAsFunction
}
\examples{
x = -3:3;
FUN = "exp( 3 * x^2 + 2 * x + 1)";
myFunction = castStringAsFunction (  FUN, x );
myFunction;
myFunction(x);

}
