% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-html.R
\name{castStringAsHTML}
\alias{castStringAsHTML}
\title{castStringAsHTML}
\usage{
castStringAsHTML(str)
}
\arguments{
\item{str}{string that is HTML code}
}
\value{
same string with appended attribute
}
\description{
RStudio/Knitr doesn't play nice with external HTML, this may fix it?
}
