% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{charAt}
\alias{charAt}
\title{charAt}
\usage{
charAt(str, idx)
}
\arguments{
\item{str}{String}

\item{idx}{position to get character}
}
\value{
single character
}
\description{
Get the character of a string at position [idx]
}
\examples{

charAt("Alex", 2);
charAt(c("Hello","there","Alex"), 2);
charAt("Alex", 8);
charAt("Alexander", 8);

}
