% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{charCode}
\alias{charCode}
\title{charCode}
\usage{
charCode(svec)
}
\arguments{
\item{svec}{A vector of characters}
}
\value{
ASCII character code for each character
}
\description{
charCode
}
\examples{

s = "Alexander"; svec = strsplit(s,"",fixed=TRUE)[[1]];
charCode(svec);

}
