% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{charCodeAt}
\alias{charCodeAt}
\title{charCodeAt}
\usage{
charCodeAt(str, idx)
}
\arguments{
\item{str}{String}

\item{idx}{position to get character}
}
\value{

}
\description{
Get the ASCII character code of a string at position [idx]
}
\examples{

charCodeAt("Alex", 2);
charCodeAt(c("Hello","there","Alex"), 2);
charCodeAt("Alex", 8);
charCodeAt("Alexander", 8);

}
