% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{checkHEX}
\alias{checkHEX}
\title{checkHEX}
\usage{
checkHEX(hex)
}
\arguments{
\item{hex}{the hex form of the color}
}
\value{
cleansed hex form of the color
}
\description{
This checks a hexadecimal input
}
\examples{
checkHEX("ABC");  # returns an updated 3-digit form
checkHEX("ABCDEF");
checkHEX("123456");

checkHEX("red"); # looks this up
checkHEX("alex"); # can't find anything in the lookup

}
