% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{checkRGB}
\alias{checkRGB}
\title{checkRGB}
\usage{
checkRGB(rgb)
}
\arguments{
\item{rgb}{the rgb input}
}
\value{
cleansed rgb
}
\description{
This checks a rgb input and replaces a bad "hex"
}
\examples{

checkRGB("red");
checkRGB(hex2rgb("red"));
checkRGB(hex2rgb("#F00"));

}
