% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-parse.R
\name{cleanup.local}
\alias{cleanup.local}
\title{cleanup.local}
\usage{
cleanup.local(myfile, append = "")
}
\arguments{
\item{myfile}{The original filename, may be a URL with ? = & symbols}
}
\value{
A windoze-friendly filename
}
\description{
cleanup.local
}
\examples{
cleanup.local("index.html?endYear=1920&amount=1000000");

}
