% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{cleanup.url}
\alias{cleanup.url}
\title{cleanup.url}
\usage{
cleanup.url(url)
}
\arguments{
\item{url}{URL to be cleansed}
}
\value{
cleansed URL
}
\description{
cleanup.url
}
\examples{
cleanup.url("https://www.myprofiletraits.com//");
cleanup.url("https:/www.mshaffer.com//arizona//");

}
