% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{cleanupRGB}
\alias{cleanupRGB}
\title{cleanupRGB}
\usage{
cleanupRGB(rgb)
}
\arguments{
\item{rgb}{matrix of rgb values}
}
\value{
matrix of rgb values on 255 scale
}
\description{
cleanupRGB
}
\examples{

rgbs = hex2rgb("#F00", "pink");
cleanupRGB(rgbs);
}
