% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.buildTable}
\alias{color.buildTable}
\title{color.buildTable}
\usage{
color.buildTable(colvec = NULL, key = NULL, save.key = NULL)
}
\arguments{
\item{colvec}{This is the colors we wish to include in our table
(a subset of the function ?colors)}

\item{key}{This is where we cache the listname tied to colvec}

\item{save.key}{This is where we cache the output dataframe generated}
}
\value{
dataframe with color (name), hex.color, r, g, b
}
\description{
color.buildTable
}
\examples{
# color.buildTable();
}
