% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.buildWheel}
\alias{color.buildWheel}
\title{color.buildWheel}
\usage{
color.buildWheel(rgb, wheel.steps = 12, find.names = FALSE)
}
\arguments{
\item{find.names}{}
}
\value{

}
\description{
color.buildWheel
}
