% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.chromatics}
\alias{color.chromatics}
\title{color.chromatics}
\usage{
color.chromatics(rgb, n = 12, save.key = NULL)
}
\arguments{
\item{save.key}{}
}
\value{

}
\description{
color.chromatics
}
