% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.colorsInGradient}
\alias{color.colorsInGradient}
\title{colorsInGradient}
\usage{
color.colorsInGradient(n, colvec = c("red", "royalblue"), alpha = FALSE)
}
\arguments{
\item{n}{Number of colors to return}

\item{colvec}{Vector of color names "red" or RGB "#FF0000" or RGBa "#4169E1FF"}

\item{alpha}{Pass transparency filter "alpha" as TRUE or FALSE}
}
\value{
vector of colors in RGB or RGBa form (depending on alpha)
}
\description{
colorsInGradient
}
\examples{

color.colorsInGradient(4, c("red", "royalblue"));
color.colorsInGradient(4, c("#FF000000", "#FF0000FF"), TRUE);  # red through alphas
color.colorsInGradient(4, c("#FF000000", "#4169E1FF"), TRUE);

}
