% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.displayColorOptions}
\alias{color.displayColorOptions}
\title{color.displayColorOptions}
\usage{
color.displayColorOptions(
  my.colors = grDevices::colors(),
  showHEX = FALSE,
  alpha = TRUE,
  xlim = c(0, 10),
  ylim = c(0, 10),
  cex = 0.75,
  ncol = 2,
  nrow = 10
)
}
\arguments{
\item{nrow}{}
}
\value{

}
\description{
color.displayColorOptions
}
