% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.findNearestName}
\alias{color.findNearestName}
\title{color.findNearestName}
\usage{
color.findNearestName(hex, how.many = 1, scale.me = TRUE, how = "cosine", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
color.findNearestName
}
