% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.nameSearch}
\alias{color.nameSearch}
\title{color.nameSearch}
\usage{
color.nameSearch(skey, colors.df = NULL, col.name = "color", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
color.nameSearch
}
