% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.plotWheel}
\alias{color.plotWheel}
\title{color.plotWheel}
\usage{
color.plotWheel(df = color.buildWheel("red"), harmony = "all")
}
\arguments{
\item{harmony}{}
}
\value{

}
\description{
color.plotWheel
}
