% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{color.setOpacity}
\alias{color.setOpacity}
\title{color.setOpacity}
\usage{
color.setOpacity(hexvec, opacity = 100)
}
\arguments{
\item{hexvec}{color in hexform}

\item{opacity}{opacity level from 0 to 100}
}
\value{
a hexform in RGBa format
}
\description{
color.setOpacity
}
\examples{
color.setOpacity("#abcdef", 0);
color.setOpacity("#abcdef", 50);
color.setOpacity("#abcdef", 100);
}
