% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-algebra.R
\name{computeCosineSimilarity}
\alias{computeCosineSimilarity}
\title{computeCosineSimilarity}
\usage{
computeCosineSimilarity(a, bs)
}
\arguments{
\item{a}{vector 'a'}

\item{bs}{vectors 'bs' (e.g., a matrix)}
}
\value{
dataframe
}
\description{
This is a multivariate (one-many) computation ...
}
\examples{
a = c(2,1,0,2,0,1,1,1); a = setAttribute("name", "a", a);
b = c(2,1,1,1,1,0,1,1);
bs = as.matrix( cbind(a, b) );
computeCosineSimilarity(a, bs);

}
