% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{computeNumericalIntegration}
\alias{computeNumericalIntegration}
\title{computeNumericalIntegration}
\usage{
computeNumericalIntegration(
  info,
  method = "string",
  FUN = "yi*1",
  tol = (.Machine$double.eps^0.25),
  skip.signs = FALSE,
  fparams = NULL,
  start = NULL,
  stop = NULL,
  verbose = FALSE,
  animatePolygons = NULL,
  showPolygons = FALSE,
  polygon.lwd = 1,
  polygon.border = NA,
  polygon.col.pos = "#4DBF4D",
  polygon.col.neg = "#BF4D4D",
  return = "result"
)
}
\arguments{
\item{return}{}
}
\value{

}
\description{
computeNumericalIntegration
}
