% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{computeNumericalIntegrationString}
\alias{computeNumericalIntegrationString}
\title{computeNumericalIntegrationString}
\usage{
computeNumericalIntegrationString(
  fstr,
  start = NULL,
  stop = NULL,
  verbose = FALSE,
  plotMe = TRUE,
  parNew = FALSE,
  showPolygons = TRUE,
  ...,
  polygon.lwd = 1,
  dxi = 0.01,
  x.domain = c(-10, 10),
  forceEven = TRUE,
  polygon.border = NA,
  polygon.col.pos = "#4DBF4D",
  polygon.col.neg = "#BF4D4D"
)
}
\arguments{
\item{polygon.col.neg}{}
}
\value{

}
\description{
computeNumericalIntegrationString
}
