% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{computeXiFromResolution}
\alias{computeXiFromResolution}
\title{computeXiFromResolution}
\usage{
computeXiFromResolution(
  x.domain,
  i.lim = c(0, 1),
  dxi = 0.01,
  forceEven = TRUE,
  oxi = 10
)
}
\arguments{
\item{oxi}{}
}
\value{

}
\description{
This is a resolution definition for numerical integration
}
