% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{count.sign.changes}
\alias{count.sign.changes}
\title{count.sign.changes}
\usage{
count.sign.changes(x, na.rm = TRUE, part = "Re")
}
\arguments{
\item{part}{}
}
\value{

}
\description{
count.sign.changes
}
