% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{createDirectoryRecursive}
\alias{createDirectoryRecursive}
\title{createDirectoryRecursive}
\usage{
createDirectoryRecursive(folder)
}
\arguments{
\item{folder}{the folder to be created}
}
\value{

}
\description{
createDirectoryRecursive
}
\examples{
createDirectoryRecursive("R:/monte/says/hi/");
}
