% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{currentInflationData}
\alias{currentInflationData}
\title{currentInflationData}
\usage{
currentInflationData(store.global = TRUE)
}
\arguments{
\item{store.global}{If TRUE, it will be accessible for inflation functions}
}
\value{
data.frame, invisible
}
\description{
This loads the most up-to-date inflation data from URL.
}
