% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-data.R
\name{data.prepare.personality}
\alias{data.prepare.personality}
\title{data.prepare.personality}
\usage{
data.prepare.personality()
}
\value{
a dataframe with the personality data loaded (via URL) and cleansed
}
\description{
See <https://github.com/MonteShaffer/humanVerse/tree/main/data/personality>
See <https://github.com/MonteShaffer/humanVerse/blob/main/data/personality/personality-raw.md>
}
\examples{
# p.df = data.prepare.personality();
}
