% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{dechex}
\alias{dechex}
\title{dechex}
\usage{
dechex(intdec, ..., n = NULL, hash = FALSE)
}
\arguments{
\item{intdec}{vector of one or more integer values}

\item{...}{vector of one or more integer values}

\item{n}{Should we prepend some zeroes, if so how many?}

\item{hash}{Should we pre..pre-pend the "#" for colors?}
}
\value{
vector of one or more hex values as a string
}
\description{
Converts a decimal to hexadecimal
}
\examples{
dechex(123,255,50, n=2, hash=FALSE);
dechex(16581375,12581375,50, n=6, hash=TRUE);
}
