% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-algebra.R
\name{deg2rad}
\alias{deg2rad}
\title{deg2rad}
\usage{
deg2rad(degs, ...)
}
\arguments{
\item{degs}{One or more angles in degrees}

\item{...}{One or more angles in degrees}
}
\value{
One or more angles in radians.
}
\description{
Convert angles from degrees to radians.
Similar to pracma::deg2rad however is vectorized (multivariate).
}
\examples{
deg2rad(c(1,3,34));
deg2rad(1,3,34);
deg2rad(1,3,"alex");

}
