% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{deleteLocalCacheFolder}
\alias{deleteLocalCacheFolder}
\title{deleteLocalCacheFolder}
\usage{
deleteLocalCacheFolder(folder)
}
\arguments{
\item{folder}{}
}
\value{

}
\description{
deleteLocalCacheFolder
}
