% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-algebra.R
\name{.angular.distance}
\alias{.angular.distance}
\title{.angular.distance}
\usage{
.angular.distance(a, b)
}
\arguments{
\item{a}{vector 'a'}

\item{b}{vector 'b'}
}
\value{
the angular distance
}
\description{
This is a univariate calculation
}
\details{
See <https://en.wikipedia.org/wiki/Cosine_similarity#Angular_distance_and_similarity>
}
\examples{
a = c(2,1,0,2,0,1,1,1); b = c(2,1,1,1,1,0,1,1);
.angular.distance(  a,b );

}
