% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{.color.average}
\alias{.color.average}
\title{.color.average}
\usage{
.color.average(a.hex, b.hex)
}
\arguments{
\item{a.hex}{The first color 'a' in hexcolor form}

\item{b.hex}{The second color 'b' in hexcolor form}
}
\value{
An "average" color in hexcolor form
}
\description{
This is univariate
}
\examples{
.color.average("#abcdef", "#123456");
}
