% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-algebra.R
\name{.cosine.similarity}
\alias{.cosine.similarity}
\title{.cosine.similarity}
\usage{
.cosine.similarity(a, b, method = "crossprod")
}
\arguments{
\item{a}{vector 'a'}

\item{b}{vector 'b'}

\item{method}{use 'crossprod' or less-efficient default option}
}
\value{
the cosine similarity
}
\description{
This is a univariate calculation
}
\details{
See <https://en.wikipedia.org/wiki/Cosine_similarity#Definition>
}
\examples{
a = c(2,1,0,2,0,1,1,1); b = c(2,1,1,1,1,0,1,1);
.cosine.similarity( a,b );

}
