% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-encryption.R, R/functions-md5.R
\name{.md5}
\alias{.md5}
\title{.md5}
\usage{
.md5(s)

.md5(s)
}
\arguments{
\item{s}{a single string, univariate}
}
\value{
an md5-string output

md5 hash of given string
}
\description{
This is a univariate function ...

Internal '.md5' function using the original RFC-1321 specs.
}
\details{
<http://md5.mshaffer.com/md5.js>
<http://block111.servehttp.com/js/md5.js>

Circa (2005)
}
\examples{
.md5("The quick brown fox jumps over the lazy dog");
.md5("alex 22");
.md5("monte");
.md5("alex");
.md5("The quick brown fox jumps over the lazy dog");

}
