% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-encryption.R, R/functions-md5.R
\name{.md5_out}
\alias{.md5_out}
\title{.md5_out}
\usage{
.md5_out(xb)

.md5_out(xb)
}
\arguments{
\item{xb}{vector of four integers}
}
\value{
md5 output of those integer keys

md5 hash from the vectors
}
\description{
This is unused, internal function of '.md5',
but may have some merit for educational purposes.
}
\examples{
.md5_out();
.md5_out();
}
