% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{extractKeysValuesFunctionParameters}
\alias{extractKeysValuesFunctionParameters}
\title{extractKeysValuesFunctionParameters}
\usage{
extractKeysValuesFunctionParameters(Keys, Vals, envir = .GlobalEnv)
}
\arguments{
\item{envir}{}
}
\value{

}
\description{
Globalizes variables for 'stepIntoFunction'
}
