% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{findAllIndexesWithNearestValueInVector}
\alias{findAllIndexesWithNearestValueInVector}
\title{findAllIndexesWithNearestValueInVector}
\usage{
findAllIndexesWithNearestValueInVector(x, search)
}
\arguments{
\item{x}{numeric vector}

\item{search}{single number}
}
\value{
the 'index' or 'indices' of the nearest value
}
\description{
findAllIndexesWithNearestValueInVector
}
\examples{
findAllIndexesWithNearestValueInVector( 1:9, 42); # 9 is nearest
findAllIndexesWithNearestValueInVector( sample(1:9), 42); # changes every time
}
