% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{findFrequencyValueInVector}
\alias{findFrequencyValueInVector}
\alias{howManyTimesDoesValueAppearInVector}
\alias{freqValue}
\title{findFrequencyValueInVector}
\usage{
findFrequencyValueInVector(x, value)
}
\arguments{
\item{x}{vector}
}
\value{
count, integer of frequency
}
\description{
What frequency is value(v) in x?
}
\examples{
f.val = 1:9;
findFrequencyValueInVector( f.val, 3 ); # 1, from all of them

f.val = c( rep(1, 3), rep(2:8,5), rep(9,1) );
findFrequencyValueInVector( f.val, 7 );

}
\seealso{
Other Vectors: 
\code{\link{doUnique}()},
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{findFrequencyMaximumInVector}()},
\code{\link{findFrequencyMinimumInVector}()},
\code{\link{notDuplicated}()},
\code{\link{whichMaxFreq}()},
\code{\link{whichMax}()},
\code{\link{whichMinFreq}()},
\code{\link{whichMin}()},
\code{\link{whichValue}()}
}
\concept{Vectors}
