% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-parse.R
\name{folderizeURL}
\alias{folderizeURL}
\title{folderizeURL}
\usage{
folderizeURL(url = "https://en.wikipedia.org/wiki/Columbia_Falls,_Montana")
}
\arguments{
\item{url}{}
}
\value{

}
\description{
Convert a url into a reversible folder structure.
}
