% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{function.getKeyFromString}
\alias{function.getKeyFromString}
\title{function.getKeyFromString}
\usage{
function.getKeyFromString(str)
}
\arguments{
\item{str}{}
}
\value{

}
\description{
Used in 'function.summary' ... would be nice to have 'traceforward()' function
}
