% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{function.whereIs}
\alias{function.whereIs}
\title{function.whereIs}
\usage{
function.whereIs(fns, ..., key = "local", getFullPath = TRUE)
}
\arguments{
\item{getFullPath}{}
}
\value{

}
\description{
Sometimes we include lots of functions in lots of files.
If registered with sourceMe ('indexFunctionsInFile'), we can see where.
}
