% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{get.sign}
\alias{get.sign}
\title{get.sign}
\usage{
get.sign(
  xs,
  ...,
  other = "NA",
  return = "number",
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{other}{default is NA, zero ... 0 ... may be applicable at times}

\item{return}{by default "number" otherwise "character"}

\item{tol}{tolerance of "zero"}

\item{x}{number(s)  to evaluate}
}
\value{
1 (+) if positive, -1 (-) if negative, "other" otherwise
}
\description{
get.sign
}
\examples{

get.sign(-1);
get.sign(-1, return="chr");
get.sign(1);
get.sign(1, return="character");
get.sign(0);
get.sign(0, return="char");
get.sign(0, 0);
get.sign(0, other=".", return="char");
get.sign(0, 0, other=".", return="char");

get.sign(-1*(1:10)); # variadic now works
get.sign(-1*(1:10), return="char");

get.sign(pi);
get.sign(-pi);
get.sign( sin(pi) );
get.sign( -sin(pi) );
get.sign( sin(pi), 0 );
get.sign( -sin(pi), 0 );
get.sign( sin(pi), tol = .Machine$double.eps^2 );
get.sign( -sin(pi), tol = .Machine$double.eps^2  );

}
