% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-date.R
\name{getDate}
\alias{getDate}
\title{getDate}
\usage{
getDate(how = "\%Y-\%m", when = Sys.time())
}
\arguments{
\item{how}{See options using ?as.Date for codes}

\item{when}{You can pass in a time, or let it grab current time}
}
\value{
date string
}
\description{
Wrapper for 'format' function
}
\examples{
getDate();
getDate("\%Y-\%m-\%d");
getDate("\%Y-\%m-\%d \%H:\%M:\%S");

getDate("\%Y-\%m-\%d", strptime("23mar1973", "\%d\%b\%Y") );
}
