% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{getFunctionParameters}
\alias{getFunctionParameters}
\title{getFunctionParameters}
\usage{
getFunctionParameters(global.memory = TRUE, n = 1, out.dput = FALSE)
}
\arguments{
\item{global.memory}{If TRUE, you can access values outside function (later)}

\item{out.dput}{}
}
\value{

}
\description{
TODO: recognize ... elements based on their names ...
}
