% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{getNearestIndex}
\alias{getNearestIndex}
\title{getNearestIndex}
\usage{
getNearestIndex(el, set, n = 1)
}
\arguments{
\item{el}{element}

\item{set}{set}

\item{n}{how many to return}
}
\value{
vector of 'indexes'
}
\description{
getNearestIndex
}
