% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{getNearestValue}
\alias{getNearestValue}
\title{getNearestValue}
\usage{
getNearestValue(el, set, n = 1)
}
\arguments{
\item{el}{element}

\item{set}{set}

\item{n}{how many to return}
}
\value{
vector of values
}
\description{
getNearestValue
}
