% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{getRemoteAndCache}
\alias{getRemoteAndCache}
\title{getRemoteAndCache}
\usage{
getRemoteAndCache(
  remote,
  local.file = NULL,
  tmp.folder = "/humanVerse/cache/",
  force.download = FALSE,
  verbose = FALSE,
  md5.hash = FALSE,
  append = ""
)
}
\arguments{
\item{append}{}
}
\value{

}
\description{
getRemoteAndCache
}
