% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{getSourceLocation}
\alias{getSourceLocation}
\title{getSourceLocation}
\usage{
getSourceLocation(tmp.subfolder = "/humanVerse/cache/", create = FALSE)
}
\arguments{
\item{create}{}
}
\value{

}
\description{
getSourceLocation
}
