% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-github.R
\name{github.includeFolder}
\alias{github.includeFolder}
\alias{includeGithubFolder}
\title{github.includeFolder}
\usage{
github.includeFolder(url, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
github.includeFolder
}
