% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-github.R
\name{github.installBinary}
\alias{github.installBinary}
\title{github.installBinary}
\usage{
github.installBinary(
  github.user = "MonteShaffer",
  github.repo = "humanVerse",
  github.path = "",
  pattern = "\\\\.zip$",
  github.version = "latest",
  ...
)
}
\value{

}
\description{
github.installBinary
}
