% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-github.R
\name{github.parseList}
\alias{github.parseList}
\title{github.parseList}
\usage{
github.parseList(url, force.download = FALSE)
}
\arguments{
\item{force.download}{}
}
\value{

}
\description{
github.parseList
}
