% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{hex2rgb}
\alias{hex2rgb}
\title{hex2rgb}
\usage{
hex2rgb(hexvec, ...)
}
\arguments{
\item{hexvec}{vector of one or more hexforms of color}

\item{...}{vector of one or more hexforms of color}
}
\value{
list of one or more rgb forms of color
}
\description{
Converts a hexform of color to rgb form
Input can be of RGB or RRGGBB hexadecimal form
Input can also be a named color
}
\examples{
hex2rgb("#FF0000");
hex2rgb(c("#FF0000", "#00FF00"));
hex2rgb("#FF0000", "#FFFFFF");


hex2rgb("red");
hex2rgb(c("red", "green"));
hex2rgb("red", "white");

hex2rgb("alex");

as.numeric( unlist( hex2rgb( "abcdef" ) ) );
unlist( hex2rgb( "ABC" ) );

}
