% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{hexdec}
\alias{hexdec}
\title{hexdec}
\usage{
hexdec(hexstr, ...)
}
\arguments{
\item{hexstr}{vector of one or more hex values as a string}

\item{...}{vector of one or more hex values as a string}
}
\value{
vector of one or more integer values
}
\description{
Converts a hexadecimal to decimal
}
\examples{
hexdec("FF");
hexdec("0xFFFF");
hexdec("0xFFFF", "#FF");
}
