% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-http.R
\name{http.headers}
\alias{http.headers}
\title{http.headers}
\usage{
http.headers(remote)
}
\arguments{
\item{remote}{A url}
}
\value{
a string vector [headers]
}
\description{
http.headers
}
\examples{
http.headers("https://www.mshaffer.com/");
http.headers("https://www.myprofiletraits.com/");
}
