% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-http.R
\name{http.headerValue}
\alias{http.headerValue}
\title{http.headerValue}
\usage{
http.headerValue(headers, search = "Content-Length:")
}
\arguments{
\item{headers}{a string vector from 'http.headers'}

\item{which}{}
}
\value{

}
\description{
http.headerValue
}
\examples{
headers = http.headers("https://www.mshaffer.com/");
http.headerValue(headers, "Content-Type");
http.headerValue(headers, "Server");
}
