% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-http.R
\name{http.size}
\alias{http.size}
\title{http.size}
\usage{
http.size(headers)
}
\arguments{
\item{headers}{a string vector from 'http.headers'}
}
\value{
The size of the page 'Content-Length'
}
\description{
http.size
}
\examples{
http.size(http.headers("https://www.mshaffer.com/"));
http.size(http.headers("http://www.myprofiletraits.com/"));
}
