% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-http.R
\name{http.status}
\alias{http.status}
\title{http.status}
\usage{
http.status(headers)
}
\arguments{
\item{headers}{a string vector from 'http.headers'}
}
\value{

}
\description{
http.status
}
\examples{
http.status(http.headers("https://www.mshaffer.com/"));
}
