% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{includeLocalDirectory}
\alias{includeLocalDirectory}
\title{includeLocalDirectory}
\usage{
includeLocalDirectory(path, verbose = TRUE, pattern = "[.][RrSsQq]$", ...)
}
\arguments{
\item{verbose}{Display details regarding includes}

\item{pattern}{This is the extension of files to be sourced ( .R )}

\item{...}{Pass extra parameters to the source function}

\item{directory}{The local directory you will scan}
}
\value{

}
\description{
includeLocalDirectory
}
