% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{includeLocalFiles}
\alias{includeLocalFiles}
\title{includeLocalFiles}
\usage{
includeLocalFiles(files, ...)
}
\arguments{
\item{files}{Vector of full file paths}

\item{...}{Pass extra parameters to the source function}
}
\value{

}
\description{
This loops through an "array" of files and sources them.
}
