% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{includeRemoteFiles}
\alias{includeRemoteFiles}
\title{includeRemoteFiles}
\usage{
includeRemoteFiles(urls, verbose = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
includeRemoteFiles
}
