% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{inflation.lookupDollar}
\alias{inflation.lookupDollar}
\alias{lookupInflationDollar}
\title{inflation.lookupDollar}
\usage{
inflation.lookupDollar(year, idf = .GlobalEnv$.humanVerse[["inflation"]])
}
\arguments{
\item{year}{numeric, 4-digit year   (1920 - 2020)}

\item{idf}{inflation data frame}
}
\value{
numeric dollar from idf table for that year
}
\description{
inflation.lookupDollar
}
\examples{
# loadInflationData();  # does lookup to create ratios
# lookupInflationDollar( 1943 ); # $  865,000
# lookupInflationDollar( 2000 ); # $8,610,000

}
\seealso{
Other Inflation: 
\code{\link{inflation.adjustDollar}()},
\code{\link{inflation.standardizeDollarsInDataFrame}()}
}
\concept{Inflation}
