% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{inflation.standardizeDollarsInDataFrame}
\alias{inflation.standardizeDollarsInDataFrame}
\alias{standardizeDollarsInDataFrame}
\title{inflation.standardizeDollarsInDataFrame}
\usage{
inflation.standardizeDollarsInDataFrame(
  df,
  anchor.year,
  dollar.source,
  year.source,
  dollar.out,
  idf = .GlobalEnv$.humanVerse[["inflation"]]
)
}
\arguments{
\item{df}{dataframe containing dollar.source and year.source}

\item{anchor.year}{base year to convert all dollars to}

\item{dollar.source}{column name (in df) with raw dollars}

\item{year.source}{column name (in df) with 4-digit years (1920 - 2020)}

\item{dollar.out}{new column name (in df) to be created}

\item{idf}{inflation data frame}
}
\value{
dataframe, updated
}
\description{
inflation.standardizeDollarsInDataFrame
}
\examples{
# loadInflationData();
# todo once I get Will/Denzel data ...
}
\seealso{
Other Inflation: 
\code{\link{inflation.adjustDollar}()},
\code{\link{inflation.lookupDollar}()}
}
\concept{Inflation}
