% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initColorMemory}
\alias{initColorMemory}
\title{initColorMemory}
\usage{
initColorMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
initColorMemory
}
