% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initFunctionMemory}
\alias{initFunctionMemory}
\title{initFunctionMemory}
\usage{
initFunctionMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
initFunctionMemory
}
