% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initInflationMemory}
\alias{initInflationMemory}
\title{initInflationMemory}
\usage{
initInflationMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
initInflationMemory
}
