% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initPathMemory}
\alias{initPathMemory}
\title{initPathMemory}
\usage{
initPathMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
initPathMemory
}
