% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-random.R
\name{initSeed}
\alias{initSeed}
\title{initSeed}
\usage{
initSeed(...)
}
\arguments{
\item{...}{Parameters that can be passed onto the 'rand' function}
}
\value{
A single integer
}
\description{
initSeed
}
\examples{
initSeed();
initSeed(1, 10^5, method="floor");
}
