% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initSeedMemory}
\alias{initSeedMemory}
\title{initSeedMemory}
\usage{
initSeedMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{purge.memory}{If TRUE, this memory is erased and reset.}
}
\value{
NULL (nothing)
}
\description{
initSeedMemory
}
\examples{
initSeedMemory();
setSeed(); getSeed(); initSeedMemory(purge.memory = TRUE); getSeed();
}
