% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-memory.R
\name{initSystemMemory}
\alias{initSystemMemory}
\title{initSystemMemory}
\usage{
initSystemMemory(purge.memory = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
initSystemMemory
}
