% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{is.negative}
\alias{is.negative}
\title{is.negative}
\usage{
is.negative(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{number (and vector) to evaluate}

\item{tol}{tolerance of "zero"}
}
\value{
TRUE OR FALSE
}
\description{
is.negative
}
\examples{

is.negative(1);
is.negative(0);
is.negative(-1);
is.negative(c(-1*1:5,1:5));

}
