% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{is.positive}
\alias{is.positive}
\title{is.positive}
\usage{
is.positive(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{number (and vector) to evaluate}

\item{tol}{tolerance of "zero"}
}
\value{
TRUE OR FALSE
}
\description{
is.positive
}
\examples{

is.positive(1);
is.positive(0);
is.positive(-1);
is.positive(c(-1*1:5,1:5));

}
