% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{is.square.matrix}
\alias{is.square.matrix}
\title{is.square.matrix}
\usage{
is.square.matrix(A)
}
\arguments{
\item{A}{input matrix}
}
\value{
NA if not a matrix, TRUE/FALSE otherwise
}
\description{
is.square.matrix
}
