% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{is.whole.number}
\alias{is.whole.number}
\title{is.whole.number}
\usage{
is.whole.number(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{number (and vector) to evaluate}

\item{tol}{tolerance of "zero"}
}
\value{
TRUE OR FALSE
}
\description{
`is.integer` doesn't operate as you would expect, this does
}
\examples{

is.whole.number(1);
is.whole.number(1.1);

is.whole.number(0);
is.whole.number(0.1);

is.whole.number(-1);
is.whole.number(-1.1);

is.whole.number(rnorm(5));
is.whole.number(rpois(5,1));

}
