% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{lastChar}
\alias{lastChar}
\title{lastChar}
\usage{
lastChar(str, trim = TRUE)
}
\arguments{
\item{str}{String}

\item{trim}{should the string be trimmed first}
}
\value{
single character
}
\description{
Get the last character of a string
}
\examples{

lastChar("Alex");
lastChar(c("Hello","there","Alex"));
lastChar("Sasha");
lastChar("Alexander");
 
}
