% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-system.R
\name{list.element.exists}
\alias{list.element.exists}
\title{list.element.exists}
\usage{
list.element.exists(mylist, myelement)
}
\arguments{
\item{listvar}{Variable Name of list}

\item{element}{Element in list to check ... can be number or string}
}
\value{
TRUE or FALSE
}
\description{
Traps error in tryCatch for non-existent list which `exists` rejects
}
\examples{

list.element.exists(some.random.nonexistent.element, "base");

my.list = list("a"=42, "b"=NULL);
  list.element.exists(my.list, "a");  # technically [["a"]] and [[1]] works in R
  list.element.exists(my.list, "b");  # null, but exists by name
  list.element.exists(my.list, "c");
  list.element.exists(my.list, 1);    # technically [["a"]] and [[1]] works in R
  list.element.exists(my.list, 2);   # it is technically NULL, but exists by name
  list.element.exists(my.list, 99);  # you shouldn't be mixing numeric / string ?

my.list = list(); my.list[[1]] = "hi"; my.list[[3]] = "alex";
  list.element.exists(my.list, 1);
  list.element.exists(my.list, 2);  # technically R may set this equal to NULL
  list.element.exists(my.list, 3);
  list.element.exists(my.list, 99);

}
