% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.addSuperNode}
\alias{matrix.addSuperNode}
\title{matrix.addSuperNode}
\usage{
matrix.addSuperNode(A, name = "=SUPER=")
}
\arguments{
\item{name}{}
}
\value{

}
\description{
matrix.addSuperNode
}
