% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.checkSquare}
\alias{matrix.checkSquare}
\alias{is.square.matrix}
\alias{matrix.is.square}
\title{matrix.checkSquare}
\usage{
matrix.checkSquare(A)
}
\arguments{
\item{A}{input matrix}
}
\value{
NA if not a matrix, TRUE/FALSE otherwise
}
\description{
matrix.checkSquare
}
