% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.computeEigenRank}
\alias{matrix.computeEigenRank}
\title{matrix.computeEigenRank}
\usage{
matrix.computeEigenRank(A, method = "linear")
}
\arguments{
\item{...}{}
}
\value{

}
\description{
matrix.computeEigenRank
}
