% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.convertMatrixToAdjacency}
\alias{matrix.convertMatrixToAdjacency}
\title{matrix.convertMatrixToAdjacency}
\usage{
matrix.convertMatrixToAdjacency(
  A,
  removeDiagonal = TRUE,
  scaleNegatives = TRUE,
  new.range = c(0, 1)
)
}
\arguments{
\item{new.range}{}
}
\value{

}
\description{
matrix.convertMatrixToAdjacency
}
