% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.pow}
\alias{matrix.pow}
\alias{powerMatrix}
\alias{matrixPower}
\title{matrix.pow}
\usage{
matrix.pow(A, pow)
}
\arguments{
\item{A}{a square matrix A}

\item{pow}{an integer power}
}
\value{
an updated square matrix A^pow
}
\description{
Raise a matrix 'A' to a power 'pow' ... works for negatives if the inverse exists
}
