% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.powerConvergence}
\alias{matrix.powerConvergence}
\alias{matrix.matrixPowerConvergence}
\title{matrix.powerConvergence}
\usage{
matrix.powerConvergence(A, tol = 1e-04)
}
\value{

}
\description{
matrix.powerConvergence
}
