% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.rank}
\alias{matrix.rank}
\alias{rankMatrix}
\alias{matrixRank}
\title{matrix.rank}
\usage{
matrix.rank(A, ...)
}
\arguments{
\item{A}{given matrix}

\item{...}{Optional parameters to pass-through}
}
\value{

}
\description{
matrix.rank
}
