% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.removeSuperNode}
\alias{matrix.removeSuperNode}
\title{matrix.removeSuperNode}
\usage{
matrix.removeSuperNode(A)
}
\arguments{
\item{A}{}
}
\value{

}
\description{
matrix.removeSuperNode
}
