% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.rotate270}
\alias{matrix.rotate270}
\alias{rotateMatrix270}
\alias{matrixRotate270}
\title{matrix.rotate270}
\usage{
matrix.rotate270(x, clockwise = TRUE)
}
\arguments{
\item{x}{matrix}

\item{clockwise}{direction of rotation, default is clockwise}
}
\value{
matrix, updated
}
\description{
matrix.rotate270
}
\examples{
m = as.matrix(structure(c(1, 0, 4, 0, 3, 0, 2, 0, 5), .Dim = c(3L, 3L)));
rotateMatrix270(m);

rotateMatrix270(m,FALSE);
rotateMatrix90(m); # equivalent

}
