% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.sortAdjacencyMatrix}
\alias{matrix.sortAdjacencyMatrix}
\title{matrix.sortAdjacencyMatrix}
\usage{
matrix.sortAdjacencyMatrix(A)
}
\arguments{
\item{A}{}
}
\value{

}
\description{
matrix.sortAdjacencyMatrix
}
