% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.trace}
\alias{matrix.trace}
\alias{traceMatrix}
\alias{matrixTrace}
\title{matrix.trace}
\usage{
matrix.trace(square)
}
\arguments{
\item{square}{a square matrix (n x n)}
}
\value{
numeric value for trace(matrix)
}
\description{
matrix.trace
}
\examples{
m = as.matrix(structure(c(1, 0, 4, 0, 3, 0, 2, 0, 5), .Dim = c(3L, 3L)));
matrixTrace(m);#'
}
