% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrix.transpose}
\alias{matrix.transpose}
\alias{transposeMatrix}
\alias{matrixTranspose}
\title{matrix.transpose}
\usage{
matrix.transpose(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
matrix, updated
}
\description{
matrix.transpose
}
\examples{
m = as.matrix(structure(c(1, 0, 4, 0, 3, 0, 2, 0, 5), .Dim = c(3L, 3L)));
transposeMatrix(m);

}
