% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrixPower}
\alias{matrixPower}
\alias{powerMatrix}
\title{matrixPower}
\usage{
matrixPower(A, pow)
}
\arguments{
\item{A}{a square matrix A}

\item{pow}{an integer power}
}
\value{
an updated square matrix A^pow
}
\description{
Raise a matrix 'A' to a power 'pow' ... works for negatives if the inverse exists
}
