% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{matrixRank}
\alias{matrixRank}
\alias{rankMatrix}
\title{matrixRank}
\usage{
matrixRank(A, ...)
}
\arguments{
\item{A}{given matrix}

\item{...}{Optional parameters to pass-through}
}
\value{

}
\description{
matrixRank
}
