% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-md5.R
\name{md5.object}
\alias{md5.object}
\title{md5.object}
\usage{
md5.object(obj, sep = "-")
}
\arguments{
\item{obj}{An object list}

\item{sep}{How you want to separate elements when you unlist}
}
\value{
An md5 hash of the object, useful for hashing parameters
passed into a function and caching the result
}
\description{
md5.object
}
\examples{
md5.object( list("alex", "is", 5) );
}
