% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-message.R
\name{message.stop}
\alias{message.stop}
\title{message.stop}
\usage{
message.stop(..., collapse = NULL, recycle0 = FALSE, pre = TRUE)
}
\arguments{
\item{...}{string elements}

\item{pre}{}
}
\value{

}
\description{
This combines 'stop' and 'paste0'
}
