% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mysql.R
\name{mysql.parseTemplateKeysVals}
\alias{mysql.parseTemplateKeysVals}
\title{mysql.parseTemplateKeysVals}
\usage{
mysql.parseTemplateKeysVals(sql, keys, vals)
}
\arguments{
\item{sql}{character string as sql.template}

\item{keys}{keys to replace in the sql.template}

\item{vals}{values to replace keys in the sql.template}
}
\value{
updated sql character string
}
\description{
mysql.parseTemplateKeysVals
}
\examples{
sql.template = "SELECT * FROM {tablename} WHERE zipcode = '{zipcode}';";
        keys = c("tablename", "zipcode");
        vals = c("zipcodes", 99163);
parseTemplateSQL(sql.template, keys, vals);

}
