% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{parseNumericalFunctionString}
\alias{parseNumericalFunctionString}
\title{parseNumericalFunctionString}
\usage{
parseNumericalFunctionString(fstr = "normal: -1, 1")
}
\arguments{
\item{fstr}{}
}
\value{

}
\description{
parseNumericalFunctionString
}
