% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-integrate.R
\name{parseTextToListOfValues}
\alias{parseTextToListOfValues}
\title{parseTextToListOfValues}
\usage{
parseTextToListOfValues(xp)
}
\arguments{
\item{xp}{}
}
\value{

}
\description{
parseTextToListOfValues
}
