% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.R
\name{plot.myHistogram}
\alias{plot.myHistogram}
\title{plot.myHistogram}
\usage{
\method{plot}{myHistogram}(
  x,
  myMain = "",
  steps.z = 1/3,
  range.z = c(-3, 3),
  ylim = c(0, 0.4),
  background.color = "#981e32",
  background.opacity = 50,
  overlay.color = "#717171",
  overlay.opacity = 50,
  do.density = TRUE,
  density.color = "#000080",
  density.opacity = 50,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
plot.myHistogram
}
