% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.R
\name{plot.myTukeyPlot}
\alias{plot.myTukeyPlot}
\title{plot.myTukeyPlot}
\usage{
\method{plot}{myTukeyPlot}(
  x.list,
  heresy = FALSE,
  sort = TRUE,
  margin = 0.25,
  border = 10,
  common.opacity = 66,
  color.bg = "#eeeeee",
  color.field = "white",
  color.hash = "#cccccc"
)
}
\arguments{
\item{color.hash}{}
}
\value{

}
\description{
plot.myTukeyPlot
}
