% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.R
\name{plot.t.test}
\alias{plot.t.test}
\title{plot.t.test}
\usage{
\method{plot}{t.test}(
  x,
  y = NULL,
  ...,
  alpha = 0.05,
  truncate.t = 4,
  plotHistogram = FALSE,
  x.lim = c(-4, 4),
  background.color = "#4DBF4D",
  background.opacity = 80,
  overlay.color = "#4D4DBF",
  overlay.opacity = 80,
  arrow.color = "#BF4D4D",
  arrow.opacity = 80
)
}
\arguments{
\item{arrow.opacity}{}
}
\value{

}
\description{
plot.t.test
}
