% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-algebra.R
\name{rad2deg}
\alias{rad2deg}
\title{rad2deg}
\usage{
rad2deg(rads, ...)
}
\arguments{
\item{...}{One or more angles in radians.}

\item{degs}{One or more angles in radians.}
}
\value{
One or more angles in degrees.
}
\description{
Convert angles from radians to degrees.
Similar to pracma::rad2deg however is vectorized (multivariate).
}
\examples{
rad2deg(c(1,3,34));
rad2deg(1,3,34);
rad2deg(1,3,"alex");

}
