% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-random.R
\name{rand}
\alias{rand}
\title{rand}
\usage{
rand(
  min = -1 * as.integer(Sys.time()),
  max = as.integer(Sys.time()),
  n = 1,
  method = "high-low",
  sample.replace = TRUE,
  seed = NULL
)
}
\arguments{
\item{min}{By default -1*as.integer(Sys.time()), what is the minimum?  Could be negative.}

\item{max}{By default as.integer(Sys.time()), what is the maximum?}

\item{n}{By default 1, how many random elements do you want?}

\item{method}{By default "high-low" which slightly outperforms "floor" which both outperform "sample"}

\item{sample.replace}{For method "sample", will allow replace = FALSE (make certain n, min/max are comformable for this)}

\item{seed}{By default NULL, meaning we are not worried about tracking the seed here.}
}
\value{

}
\description{
Generate random integers between two values (min, max).
}
\examples{
rand();            # positive or negative integer
rand(1);           # positive integer only
rand(1, n=3);      # returns 3 positive integers
rand(1,10, n=5, method="floor");  # Uses the floor method
rand(1,10, n=5, method="sample"); # Uses the sample method (available, but why?)
rand(1,10, n=5, method="sample", sample.replace=FALSE); # min, max, n must be comformable "with replacement = FALSE"
}
