% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{readFromPipe}
\alias{readFromPipe}
\title{readFromPipe}
\usage{
readFromPipe(file, header = TRUE, quote = "", sep = "|")
}
\arguments{
\item{file}{filename and path}

\item{header}{whether or not to add a header}

\item{quote}{whether or not to add quotes}

\item{sep}{"pipe" means sep="|" but you could change}
}
\value{
a dataframe
}
\description{
This is the inverse of 'writeToPipe'
}
