% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{readRDS.url}
\alias{readRDS.url}
\title{readRDS.url}
\usage{
readRDS.url(file)
}
\arguments{
\item{file}{The file is likely a URL in this function}
}
\value{
a data object, likely a dataframe or a list
}
\description{
This wraps 'url' and 'readRDS' so they are webfriendly ...
}
