% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{readStringFromFile}
\alias{readStringFromFile}
\title{readStringFromFile}
\usage{
readStringFromFile(myFile, n = NULL, method = "readChar", source = "local")
}
\arguments{
\item{source}{}
}
\value{

}
\description{
readStringFromFile
}
