% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{removeFunnyCharacters}
\alias{removeFunnyCharacters}
\title{removeFunnyCharacters}
\usage{
removeFunnyCharacters(str, which = "alpha-numeric", ASCII = TRUE, trim = TRUE)
}
\arguments{
\item{ASCII}{}
}
\value{

}
\description{
removeFunnyCharacters
}
