% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeNAsFromDataFrame}
\alias{removeNAsFromDataFrame}
\title{removeNAsFromDataFrame}
\usage{
removeNAsFromDataFrame(df, mycols = NULL)
}
\arguments{
\item{df}{data.frame to cleanse}

\item{mycols}{cols to check for NA, by default NULL is all columns}
}
\value{
dataframe, updated
}
\description{
removeNAsFromDataFrame
}
\examples{
library(datasets);
data(iris);

df = iris;
df[1,1] = NA; df[2,2] = NA; df[3,3] = NA; df[4,4] = NA; df[5,5] = NA;

head(df);
dim(df);

ndf = removeNAsFromDataFrame(df, c("Petal.Length","Petal.Width","Species"));
head(ndf);
dim(ndf);


ndf = removeNAsFromDataFrame(df);
head(ndf);
dim(ndf);
}
