% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-system.R
\name{returnList}
\alias{returnList}
\title{returnList}
\usage{
returnList(res, unlist = TRUE)
}
\arguments{
\item{unlist}{}
}
\value{

}
\description{
If list only has one element return it.
By default unlist so it is just a vector ... (redundant with above?)
}
