% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{rgb2col}
\alias{rgb2col}
\title{rgb2col}
\usage{
rgb2col(x)
}
\arguments{
\item{x}{vector of colors}
}
\value{
vector of colors in RGB hex format
}
\description{
Reverse the built-in grDevices::col2rgb function
[ See grDevices::convertColor or grDevices::make.rgb ]
}
\examples{

rgb2col( grDevices::col2rgb("red") );
rgb2col( grDevices::col2rgb("red", alpha=TRUE) );
rgb2col( grDevices::col2rgb("#FF0000FF", alpha=TRUE) );
rgb2col( grDevices::col2rgb("#FF000033", alpha=TRUE) );

}
