% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-colors.R
\name{rgb2hex}
\alias{rgb2hex}
\title{rgb2hex}
\usage{
rgb2hex(rgb, pre = "#")
}
\arguments{
\item{rgb}{matrix of RGB values}

\item{pre}{do we want to pre-pend the hash "#" ?}
}
\value{
vector of hexform colors
}
\description{
rgb2hex
}
\examples{
rgbs = hex2rgb("#F00", "pink");
rgb2hex(rgbs);#'
}
