% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-parse.R
\name{sliceDiceContent}
\alias{sliceDiceContent}
\title{sliceDiceContent}
\usage{
sliceDiceContent(
  str,
  start = "<h2>",
  end = "</h2>",
  strip = FALSE,
  direction = "start"
)
}
\arguments{
\item{direction}{}
}
\value{

}
\description{
This takes a string and returns a substring based on some
 'cut points' defined as start and end.
}
\details{
The main engine is strsplit so there has to be a direction.
}
