% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeFromOldRangeToNew}
\alias{standardizeFromOldRangeToNew}
\title{standardizeFromOldRangeToNew}
\usage{
standardizeFromOldRangeToNew(x, oldrange = range(x), newrange = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{oldrange}{this is typically the current range of the vector [numeric vector of length 2]}

\item{newrange}{this is the new scaled range [numeric vector of length 2]}
}
\value{
numeric vector, updated
}
\description{
standardizeFromOldRangeToNew
}
\examples{
standardizeFromOldRangeToNew( 1:5, newrange=c(0,1) );  # map to [0,1]
standardizeFromOldRangeToNew( runif(20, min=-1, max=1), newrange=c(0,1) );  # map to [0,1]
standardizeFromOldRangeToNew( 1:5, newrange=c(1,7) );  # e.g., Likert-5 to Likert-7
standardizeFromOldRangeToNew( rep(1,10), newrange=c(0,1) );  #NaN ... division by zero

}
