% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plot.R
\name{stats.prepareTukeyPlot}
\alias{stats.prepareTukeyPlot}
\title{stats.prepareTukeyPlot}
\usage{
stats.prepareTukeyPlot(
  x.list,
  heresy = FALSE,
  sort.me = TRUE,
  sort.list = "ASC",
  sort.key = "median"
)
}
\arguments{
\item{sort.key}{}
}
\value{

}
\description{
stats.prepareTukeyPlot
}
