% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-functions.R
\name{stepIntoFunction}
\alias{stepIntoFunction}
\title{stepIntoFunction}
\usage{
stepIntoFunction(obj)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
The idea is to call this function so you don't have to
prepopulate default parameters.
}
