% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-parse.R
\name{stringifyLibrary}
\alias{stringifyLibrary}
\title{stringifyLibrary}
\usage{
stringifyLibrary(str)
}
\arguments{
\item{str}{}
}
\value{

}
\description{
If multiple users share a notebook (Dropbox), you can identify
 who they are by the R_LIBS_HOME or some other element
}
