% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-parse.R
\name{strip_tags}
\alias{strip_tags}
\alias{strip.tags}
\alias{striptags}
\title{strip_tags}
\usage{
strip_tags(htmlString)
}
\arguments{
\item{htmlString}{}
}
\value{

}
\description{
Remove HTML tags from a string.
}
