% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{trimMe}
\alias{trimMe}
\title{trimMe}
\usage{
trimMe(str, side = "both", method = "stringi")
}
\arguments{
\item{str}{character string to be "trimmed"}
}
\value{
updated trimmed string
}
\description{
library(stringi);
}
\examples{

trimMe( c(" Monte", " is ", "Alexander's ", "  daddy!") );

trimMe("    four   scores    and  seven      years     ");
trimMe("    four   scores    and  seven      years     ", "left");
trimMe("    four   scores    and  seven      years     ", "riGht");
trimMe("    four   scores    and  seven      years     ", "both");
trimMe("    four   scores    and  seven      years     ", "albjdskj")

trimMe("\r\n    four   scores    and  seven      years   \t\t  ");

}
