% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-print.R
\name{truncateNumeric}
\alias{truncateNumeric}
\title{truncateNumeric}
\usage{
truncateNumeric(val, digits = 3, append = "...")
}
\arguments{
\item{val}{numeric value}

\item{digits}{how many digits}

\item{append}{what to append (postpend) to demonstrate truncation}
}
\value{
stringified form of val
}
\description{
truncateNumeric
}
\examples{
truncateNumeric(3.1415926535897932384626);
truncateNumeric(3.1415926535897932384626, digits=22);
truncateNumeric('3.1415926535897932384626', digits=22);
truncateNumeric(3.1415926535897932384626, digits=9, append="");
}
