% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-number.R
\name{which.sign.changes}
\alias{which.sign.changes}
\title{which.sign.changes}
\usage{
which.sign.changes(x, na.rm = TRUE, part = "Re")
}
\arguments{
\item{part}{}
}
\value{

}
\description{
which.sign.changes
}
