% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{whichNearestValue}
\alias{whichNearestValue}
\title{whichNearestValue}
\usage{
whichNearestValue(x, search)
}
\arguments{
\item{x}{numeric vector}

\item{search}{single number}
}
\value{
The nearest value from the found 'index'
}
\description{
whichNearestValue
}
\examples{
whichNearestValue( 1:9, 42); # 9 as in the number
whichNearestValue( sample(1:9), 42); # 9 as in the number
}
