% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{whichValue}
\alias{whichValue}
\title{whichValue}
\usage{
whichValue(x, value, ...)
}
\arguments{
\item{x}{vector}

\item{value}{numeric value}

\item{...}{other elements to pass into findAllIndexesWithValueInVector (method)}
}
\value{
vector that contains the indexes of *all* v elements, not just the *first*
}
\description{
wraps which function for vectors
}
\examples{
whichValue( presidents[1:30], 87 );
whichValue( presidents[1:30], NA );  # doesn't work as I would expect
whichValue( presidents[1:30], NA, method="which" );
whichValue( presidents[1:30], NA, method="hack-a-thon" );
whichValue( presidents[1:30], NA, method="hackAth)N" );
whichValue( presidents[1:30], NA, method="NA" );
whichValue( presidents[1:30], NA, "NA" ); # works as I would expect
}
\seealso{
Other Vectors: 
\code{\link{doUnique}()},
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{findFrequencyMaximumInVector}()},
\code{\link{findFrequencyMinimumInVector}()},
\code{\link{findFrequencyValueInVector}()},
\code{\link{notDuplicated}()},
\code{\link{whichMaxFreq}()},
\code{\link{whichMax}()},
\code{\link{whichMinFreq}()},
\code{\link{whichMin}()}
}
\concept{Vectors}
