% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{writeRDS}
\alias{writeRDS}
\title{writeRDS}
\usage{
writeRDS(obj, myfile)
}
\arguments{
\item{obj}{The object to be stored}

\item{myfile}{The file to store the object}
}
\value{

}
\description{
The opposite of readRDS is writeRDS, make it so.
}
