% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{writeToPipe}
\alias{writeToPipe}
\alias{storeToPipe}
\title{writeToPipe}
\usage{
writeToPipe(df, file, header = TRUE, quote = "", sep = "|", row.names = FALSE)
}
\arguments{
\item{df}{dataframe to be stored}

\item{file}{filename and path}

\item{header}{whether or not to add a header}

\item{quote}{whether or not to add quotes}

\item{sep}{"pipe" means sep="|" but you could change}

\item{row.names}{whether or not to include row names}
}
\value{
NOTHING, it writes
}
\description{
This is the inverse of 'readFromPipe'
}
