% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{zCutOverlay}
\alias{zCutOverlay}
\title{zCutOverlay}
\usage{
zCutOverlay(z.table, steps.z = 1/2, verbose = FALSE, myColor = "blue", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
zCutOverlay
}
