% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{getSourceLocation}
\alias{getSourceLocation}
\title{getSourceLocation}
\usage{
getSourceLocation(tmp.folder = "/humanVerse/cache/")
}
\arguments{
\item{tmp.folder}{}
}
\value{

}
\description{
getSourceLocation
}
